requirements.txt
gradio==5.29.0
graphiti-core==0.11.6
python-dotenv==1.1.0
pydantic==2.11.0

neo4j==5.28.1

openai==1.82.0

PyPDF2==3.0.1
